This required a lot of conflict resolution since the two branches had diverged quite a bit.
+ Fixed the way the video filter name/friendly name is retrieved.

+++ Kinovea.ScreenManager/Metadata/Drawings/DrawingSVG.cs
@@ -159,7 +159,7 @@ namespace Kinovea.ScreenManager
         }
         public override void MoveDrawing(float dx, float dy, Keys _ModifierKeys, bool zooming)
         {
-            boundingBox.MoveAndSnap((int)dx, (int)dy, videoSize, snapMargin);
+            boundingBox.MoveAndSnap(dx, dy, videoSize, snapMargin);
         }
         public override PointF GetCopyPoint()
         {
